Some classes were improved, some bugs were fixed, unnecessary comments were removed

+++ src/main/java/org/example/services/impl/OrderItemServiceImpl.java
@@ -3,6 +3,7 @@ package org.example.services.impl;
 import jakarta.transaction.Transactional;
 import org.example.dto.OrderItemCreateDto;
 import org.example.dto.OrderItemDto;
+import org.example.enums.Role;
 import org.example.exceptions.OrderItemNotFoundException;
 import org.example.exceptions.OrderNotFoundException;
 import org.example.exceptions.ProductNotFoundException;
@@ -11,10 +12,13 @@ import org.example.mappers.OrderItemMapper;
 import org.example.models.Order;
 import org.example.models.OrderItem;
 import org.example.models.Product;
+import org.example.models.User;
 import org.example.repositories.OrderItemRepository;
 import org.example.repositories.OrderRepository;
 import org.example.repositories.ProductRepository;
+import org.example.security.SecurityUtils;
 import org.example.services.interfaces.OrderItemService;
+import org.example.services.interfaces.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -28,37 +32,43 @@ public class OrderItemServiceImpl implements OrderItemService {
     private final OrderRepository orderRepository;
     private final ProductRepository productRepository;
     private final OrderItemMapper orderItemMapper;
+    private final SecurityUtils securityUtils;
 
-    @Autowired
     public OrderItemServiceImpl(OrderItemRepository orderItemRepository,
                                 OrderRepository orderRepository,
                                 ProductRepository productRepository,
-                                OrderItemMapper orderItemMapper) {  // Внедрение маппера
+                                OrderItemMapper orderItemMapper,
+                                SecurityUtils securityUtils) {
         this.orderItemRepository = orderItemRepository;
         this.orderRepository = orderRepository;
         this.productRepository = productRepository;
-        this.orderItemMapper = orderItemMapper;  // Инициализация
+        this.orderItemMapper = orderItemMapper;
+        this.securityUtils = securityUtils;
     }
 
+    private User getCurrentUser() {
+        return securityUtils.getCurrentUser();
+    }
     @Transactional
     @Override
     public OrderItemDto createOrderItem(OrderItemCreateDto orderItemCreateDto, Long orderId) {
-        // Получаем заказ по ID
+        User currentUser = getCurrentUser();
+
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new OrderNotFoundException(ErrorMessage.ORDER_NOT_FOUND));
+        checkOrderOwnership(order, currentUser);
 
-        // Получаем продукт по ID
-        Product product = productRepository.findById(orderItemCreateDto.getProductId())
-                .orElseThrow(() -> new ProductNotFoundException(ErrorMessage.PRODUCT_NOT_FOUND));
 
-        // Создаем новый OrderItem из DTO и Product
+        Product product = getProductById(orderItemCreateDto.getProductId());
+
+
         OrderItem orderItem = orderItemMapper.toEntity(orderItemCreateDto, product);
         orderItem.setOrder(order);
 
-        // Сохраняем новый OrderItem
+
         orderItem = orderItemRepository.save(orderItem);
 
-        // Обновляем общую сумму заказа
+
         updateOrderTotalAmount(order);
 
         return orderItemMapper.toDto(orderItem);
@@ -66,6 +76,12 @@ public class OrderItemServiceImpl implements OrderItemService {
 
     @Override
     public List<OrderItemDto> getOrderItemsByOrderId(Long orderId) {
+        User currentUser = getCurrentUser();
+
+        Order order = orderRepository.findById(orderId)
+                .orElseThrow(() -> new OrderNotFoundException(ErrorMessage.ORDER_NOT_FOUND));
+
+        checkOrderOwnership(order, currentUser);
         return orderItemRepository.findByOrderId(orderId).stream()
                 .map(orderItemMapper::toDto)
                 .collect(Collectors.toList());
@@ -73,20 +89,23 @@ public class OrderItemServiceImpl implements OrderItemService {
     @Transactional
     @Override
     public OrderItemDto updateOrderItem(Long orderItemId, OrderItemCreateDto orderItemCreateDto) {
+        User currentUser = getCurrentUser();
         OrderItem orderItem = orderItemRepository.findById(orderItemId)
                 .orElseThrow(() -> new OrderItemNotFoundException(ErrorMessage.ORDER_ITEM_NOT_FOUND));
+        Order order = orderItem.getOrder();
 
-        // Получаем продукт по ID
-        Product product = productRepository.findById(orderItemCreateDto.getProductId())
-                .orElseThrow(() -> new ProductNotFoundException(ErrorMessage.PRODUCT_NOT_FOUND));
+        checkOrderOwnership(order, currentUser);
 
-        // Обновляем данные OrderItem
+
+        Product product = getProductById(orderItemCreateDto.getProductId());
+
+
         orderItemMapper.updateEntityFromCreateDto(orderItemCreateDto, orderItem, product);
 
-        // Сохраняем обновленный OrderItem
+
         orderItem = orderItemRepository.save(orderItem);
 
-        // Обновляем общую сумму заказа
+
         updateOrderTotalAmount(orderItem.getOrder());
 
         return orderItemMapper.toDto(orderItem);
@@ -103,18 +122,27 @@ public class OrderItemServiceImpl implements OrderItemService {
     @Transactional
     @Override
     public void deleteOrderItem(Long orderItemId) {
+        User currentUser = getCurrentUser();
         OrderItem orderItem = orderItemRepository.findById(orderItemId)
                 .orElseThrow(() -> new OrderItemNotFoundException(ErrorMessage.ORDER_ITEM_NOT_FOUND));
 
         Order order = orderItem.getOrder();
 
-        // Удаляем OrderItem из коллекции Order
-        order.getOrderItems().remove(orderItem);
 
-        // Удаляем OrderItem из базы данных
+        checkOrderOwnership(order, currentUser);
+
         orderItemRepository.delete(orderItem);
 
-        // Обновляем общую сумму заказа
+
         updateOrderTotalAmount(order);
+    }
+    private Product getProductById(Long productId) {
+        return productRepository.findById(productId)
+                .orElseThrow(() -> new ProductNotFoundException(ErrorMessage.PRODUCT_NOT_FOUND));
+    }
+    private void checkOrderOwnership(Order order, User user) {
+        if (!order.getUser().getId().equals(user.getId()) && !user.getRole().equals(Role.ADMIN)) {
+            throw new org.springframework.security.access.AccessDeniedException(ErrorMessage.ACCESS_DENIED);
+        }
     }
 }
